requirements.txt
kivy
requests
pyserial
ecdsa
python3
